image/src/search_requirements.txt
pydantic==2.10.6
psycopg2-binary==2.9.10